+++ rita_lab/Cargo.toml
@@ -20,6 +20,11 @@ license = "MIT"
 rita = "0.1.7"
 vertex_clustering = "0.1.1"
 csv = "1.2.2"
+rfd = "0.15.3"
+
+log = "0.4"
+rita_test_utils = { version = "0.1", path = "../rita_test_utils" }
+
 egui = "0.27.2"
 egui_plot = { version = "0.27.2", default-features = false }
 eframe = { version = "0.27.2", default-features = false, features = [
@@ -28,17 +33,13 @@ eframe = { version = "0.27.2", default-features = false, features = [
     "glow",          # Use the glow rendering backend. Alternative: "wgpu".
     "persistence",   # Enable restoring app state when restarting the app.
 ] }
-rfd = "0.14.1"
-log = "0.4"
-rand_distr = "0.4.3"
 # You only need serde if you want app persistence:
 serde = { version = "1", features = ["derive"] }
-rand = "0.8.5"
 
 # native:
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-env_logger = "0.11.3"
-futures = "0.3.28"
+env_logger = "0.11.8"
+futures = "0.3.31"
 
 # web:
 [target.'cfg(target_arch = "wasm32")'.dependencies]
@@ -51,14 +52,3 @@ wasm-bindgen-futures = "0.4"
 # Optimize all dependencies even in debug builds:
 #[profile.dev.package."*"]
 #opt-level = 2
-
-
-#[patch.crates-io]
-
-# If you want to use the bleeding edge version of egui and eframe:
-# egui = { git = "https://github.com/emilk/egui", branch = "master" }
-# eframe = { git = "https://github.com/emilk/egui", branch = "master" }
-
-# If you fork https://github.com/emilk/egui you can test with:
-# egui = { path = "../egui/crates/egui" }
-# eframe = { path = "../egui/crates/eframe" }
